requirements.txt
playwright
pyyaml
requests
httpx

# https://github.com/kkroening/ffmpeg-python
ffmpeg-python

# for debugging and tracing
snoop
# for cli parsing
fire